Added comments and functions for computing and caching inverse of a square invertible matrix
Since matrix inversion is a costly computation, this pair of functions computes and caches the inverse of a matrix. makeCacheMatrix creates a "special" matrix object that can cache its inverse. The cacheSolve function computes the inverse of the special "matrix" returned by makeCacheMatrix. If the inverse has already been calculated, then the cacheSolve should retrieve the inverse from the cache.  I added the comments and the functions to this file.